@cijothomas 2022-10-26
> Nice! It's the little things that count

All these, so you won't complain when I add more View stuffs and drop perf 🏃